[NFC] Extract dispatch region heuristic into separate file
