+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/BUILD
@@ -44,6 +44,7 @@ iree_compiler_cc_library(
         "DetachElementwiseFromNamedOps.cpp",
         "DispatchLinalgOnTensors.cpp",
         "DispatchLinalgOnTensorsViaRegionOps.cpp",
+        "DispatchRegionHeuristic.cpp",
         "DispatchWithTransformDialect.cpp",
         "DumpDispatchGraph.cpp",
         "ExpandTensorShapes.cpp",
@@ -70,6 +71,7 @@ iree_compiler_cc_library(
     hdrs = [
         "ConvertRegionToWorkgroups.h",
         "DispatchLinalgOnTensors.h",
+        "DispatchRegionHeuristic.h",
         "Passes.h",
         "Passes.h.inc",
         "RegionOpUtils.h",
