+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/CMakeLists.txt
@@ -25,6 +25,7 @@ iree_cc_library(
   HDRS
     "ConvertRegionToWorkgroups.h"
     "DispatchLinalgOnTensors.h"
+    "DispatchRegionHeuristic.h"
     "Passes.h"
     "Passes.h.inc"
     "RegionOpUtils.h"
@@ -42,6 +43,7 @@ iree_cc_library(
     "DetachElementwiseFromNamedOps.cpp"
     "DispatchLinalgOnTensors.cpp"
     "DispatchLinalgOnTensorsViaRegionOps.cpp"
+    "DispatchRegionHeuristic.cpp"
     "DispatchWithTransformDialect.cpp"
     "DumpDispatchGraph.cpp"
     "ExpandTensorShapes.cpp"
